auth

+++ CosmosTrainingAPi/Controllers/PostController.cs
@@ -9,32 +9,39 @@ namespace CosmosTrainingAPi.Controllers
     [ApiController]
     public class PostController : ControllerBase
     {
-        private ICosmosDBService _cosmosDBService;
+        private readonly ICosmosDBService _cosmosDBService;
+        private readonly IUserService _userService;
 
-        public PostController(ICosmosDBService cosmosDBService)
+        public PostController(ICosmosDBService cosmosDBService, IUserService userService)
         {
             _cosmosDBService = cosmosDBService;
+            _userService = userService;
         }
 
-        [HttpPost]
+        [HttpPost, Authorize(Roles = "User")]
         public async Task<ActionResult<string>> addPost(Post post)
         {
             string responce = await _cosmosDBService.createPost(post);
             return Ok(responce);
         }
 
-        [HttpGet, Authorize]
+        [HttpGet]
         public async Task<ActionResult<List<UserDTO>>> GetAllposts()
         {
             var responce = await _cosmosDBService.GetAllposts();
             return Ok(responce);
         }
 
-        [HttpDelete]
-        public async Task<ActionResult<List<UserDTO>>> DeletePost(DeletePost post)
+        [HttpDelete, Authorize(Roles = "User")]
+        public async Task<ActionResult<string>> DeletePost(DeletePost post)
         {
-            var responce = await _cosmosDBService.DeletePost(post);
-            return Ok(responce);
+            if (_userService.parseRoleFromToken().Contains("Admin") || _userService.parseNameFromToken() == post.Username)
+            {
+                var responce = await _cosmosDBService.DeletePost(post);
+                return Ok(responce);
+            }
+            else
+                return Unauthorized("");
         }
     }
 }
